Split app container composition
