+++ apps/server/vibesensor/app/composition/__init__.py
@@ -1,0 +1,1 @@
+"""Subsystem-specific factories for the application composition root."""
